fix(web): move two unit-tested pure helpers out of their component files
valueFontSize and splitInlinedHistory were exported from .tsx files, so importing
them in a test pulled lucide-react's CJS bundle into the process, which fails to
resolve React in a clean environment: green locally, red on a fresh checkout. The new
CI workflow surfaced it — bun test had never run in CI before, since release.yml only
fires on a tag.

Both are pure functions with their own tests, so lib/ is where the project's own
conventions already put them. No behaviour change; the web build is unaffected.
